requirements.txt
Misskey.py
websockets
python-dotenv
google-genai